+++ @navikt/core/react/src/data/stories/Data.wofw.stories.tsx
@@ -72,10 +72,10 @@ export const ExampleWithoutTanstack: Story = {
     /* const [columnPinning, setColumnPinning] = useState<{
       left: string[];
       right: string[];
-    }>({ left: [], right: [] });
+    }>({ left: [], right: [] });*/
     const [columnSizes, setColumnSizes] = useState<{ [key: string]: number }>(
       {},
-    ); */
+    );
 
     const globalFilterLower = deferredFilterString.toLowerCase();
     // Perf: Memoize data to avoid rerendering table body when unrelated state changes (the filtering itself is not expensive)
@@ -92,19 +92,19 @@ export const ExampleWithoutTanstack: Story = {
     );
 
     function resizeHandler(event: React.MouseEvent<HTMLButtonElement>) {
-      /* const startX = event.clientX; */
+      const startX = event.clientX;
       const th = (event.target as HTMLElement).closest(
         "th",
       ) as HTMLTableCellElement;
-      /* const startWidth = th.offsetWidth; */
-      function onMouseMove() {
-        /* const newWidth = startWidth + (e.clientX - startX); */
+      const startWidth = th.offsetWidth;
+      function onMouseMove(e: MouseEvent) {
+        const newWidth = startWidth + (e.clientX - startX);
         const colKey = th.dataset.key;
         if (!colKey) return;
-        /* setColumnSizes((prev) => ({
+        setColumnSizes((prev) => ({
           ...prev,
           [colKey]: newWidth,
-        })); */
+        }));
       }
       function onMouseUp() {
         document.removeEventListener("mousemove", onMouseMove);
@@ -113,8 +113,6 @@ export const ExampleWithoutTanstack: Story = {
       document.addEventListener("mousemove", onMouseMove);
       document.addEventListener("mouseup", onMouseUp);
     }
-
-    /* console.log(columnSizes); */
 
     return (
       <VStack gap="space-16">
@@ -132,7 +130,7 @@ export const ExampleWithoutTanstack: Story = {
                 return (
                   <DataTable.Th
                     key={column.header}
-                    /* size={columnSizes[column.accessorKey] ?? 150} */
+                    size={columnSizes[column.accessorKey] ?? 150}
                     //style={{ width: `var(--header-${header.id}-size)` }}
                     resizeHandler={resizeHandler}
                     data-key={column.accessorKey}
@@ -184,7 +182,4 @@ const TableBody = ({ data }: { data: PersonInfo[] }) => (
   </DataTable.Tbody>
 );
 
-const MemoizedTableBody = React.memo(
-  TableBody,
-  (prev, next) => prev.data === next.data,
-) as typeof TableBody;
+const MemoizedTableBody = React.memo(TableBody) as typeof TableBody;