+++ @navikt/core/react/src/data/stories/DataTableProfiler.tsx
@@ -56,6 +56,12 @@ function rollingLatestValues(
   return newArr;
 }
 
+const getOpacity = (timestamp: number) => {
+  const age = Date.now() - timestamp;
+  const maxAge = 10000;
+  return Math.max(0.4, 1 - age / maxAge);
+};
+
 function ProfilerDisplay({
   getLatestData,
 }: {
@@ -96,12 +102,6 @@ function ProfilerDisplay({
     }, 500);
     return () => clearInterval(interval);
   }, [open]);
-
-  const getOpacity = (timestamp: number) => {
-    const age = Date.now() - timestamp;
-    const maxAge = 10000;
-    return Math.max(0.4, 1 - age / maxAge);
-  };
 
   return (
     <VStack asChild>