+++ @navikt/core/react/src/form/file-upload/helpers/file-type-checker.ts
@@ -1,4 +1,4 @@
-import { FileItem } from "../Item.types";
+import { FileItem } from "../item/FileUploadItem.types";
 
 export const isNativeFile = (fileItem: FileItem): fileItem is File =>
   "lastModified" in fileItem;