+++ @navikt/core/react/src/form/file-upload/hooks/useFileUpload.ts
@@ -1,10 +1,12 @@
 import { useRef } from "react";
-import { useMergeRefs } from "../../utils/hooks";
-import { FileUploadBaseProps } from "./FileUpload.types";
-import { validateFiles } from "./utils/validate-files";
+import { useMergeRefs } from "../../../utils/hooks";
+import { FileUploadBaseProps } from "../FileUpload.types";
+import { validateFiles } from "../helpers/validate-files";
 
-export interface UseFileUploadProps
-  extends Omit<FileUploadBaseProps, "fileLimit"> {
+export interface UseFileUploadProps extends Omit<
+  FileUploadBaseProps,
+  "fileLimit"
+> {
   ref: React.ForwardedRef<HTMLInputElement>;
   disabled?: boolean;
 }