+++ @navikt/core/react/src/utils/i18n/locales/nb.ts
@@ -139,4 +139,7 @@ export default {
       reset: "Tilbakestill tidsperspektiv",
     },
   },
+  Tooltip: {
+    shortcutSeparator: "eller",
+  },
 } satisfies TranslationMap;