+++ aksel.nav.no/website/pages/templates/500/komplett.tsx
@@ -9,6 +9,7 @@ import {
   Page,
   VStack,
 } from "@navikt/ds-react";
+import { withDsExample } from "@/web/examples/withDsExample";
 import {
   Env,
   Footer,
@@ -103,7 +104,13 @@ function Example() {
 }
 
 // EXAMPLES DO NOT INCLUDE CONTENT BELOW THIS LINE
-export default Example;
+export default withDsExample(Example, {
+  theme: {
+    forcedTheme: "light",
+    switch: false,
+  },
+  variant: "fullscreen",
+});
 
 /* Storybook story */
 export const Demo = {