+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## 8.3.0
+
+### @navikt/ds-react
+
+- Tooltip: Updated prop 'keys' to allow for multiple shortcuts. ([#4604](https://github.com/navikt/aksel/pull/4604))
+
+### @navikt/ds-css
+
+- Tooltip: Updated prop 'keys' to allow for multiple shortcuts. ([#4604](https://github.com/navikt/aksel/pull/4604))
+
+* Chat: Fix avatar size when using icon from aksel-icons ([#4608](https://github.com/navikt/aksel/pull/4608))
+
 ## 8.2.2
 
 ### @navikt/ds-react